+++ include/rotatingMHD/run_time_parameters.h
@@ -1,10 +1,13 @@
 #ifndef INCLUDE_ROTATINGMHD_RUN_TIME_PARAMETERS_H_
 #define INCLUDE_ROTATINGMHD_RUN_TIME_PARAMETERS_H_
 
-#include <deal.II/base/parameter_handler.h>
+#include <rotatingMHD/basic_parameters.h>
+#include <rotatingMHD/linear_solver_parameters.h>
 #include <rotatingMHD/convergence_test.h>
-
 #include <rotatingMHD/time_discretization.h>
+
+#include <deal.II/base/parameter_handler.h>
+
 #include <memory>
 
 namespace RMHD
@@ -16,350 +19,6 @@ namespace RMHD
  */
 namespace RunTimeParameters
 {
-
-using namespace dealii;
-
-/*!
- * @brief Enumeration for the different types of problems.
- *
- * @details The different problems derive from the stencil
-  * \f[
-  * \begin{gather*}
-    \nabla \cdot \bs{v}
-    = 0, \quad
-      \pd{\bs{v}}{t} + \bs{v} \cdot (\nabla \otimes \bs{v})
-      + C_1 \bs{\Omega} \times \bs{v}
-    = -C_6 \nabla \bar{p}
-    + C_ 2\nabla^2 \bs{v}
-    - C_3 \vartheta \bs{g}
-    + C_5 (\nabla \times \bs{B}) \times \bs{B},
-    \\
-    \pd{\vartheta}{t} + \bs{v} \cdot \nabla \vartheta
-    =
-    C_4 \nabla^2 \vartheta, \\
-    \nabla \cdot \bs{B} = 0, \quad
-    \pd{\bs{B}}{t} = \nabla \times ( \bs{v} \times \bs{B} ) + C_5 \nabla \times ( \nabla \times \bs{B} ),    \qquad \qquad \forall (\bs{x}, t) \in \Omega \times [0, T],
-    \end{gather*}
-  * \f]
-  * where \f$ \bs{v} \f$ is the velocity,
-  * \f$ \bar{p} \f$ the modified pressure,
-  * \f$ \vartheta \f$ the temperature,
-  * \f$ \bs{B} \f$ the magnetic field and
-  * \f$ {C_i} \f$ the dimensionless coefficients. These coefficients
-  * are given by a combination of dimensionless numbers depending of
-  * the problem type.
-  * @note For a definition of the pertinent dimensionless numbers see
-  * @ref DimensionlessNumbers
- */
-enum class ProblemType
-{
-  /*!
-   * @brief The Navier-Stokes equations in an intertial frame of reference.
-   * @details Given by
-  * \f[
-  * \begin{gather*}
-    \nabla \cdot \bs{v}
-    = 0, \quad
-      \pd{\bs{v}}{t} + \bs{v} \cdot (\nabla \otimes \bs{v})
-    = - \nabla \bar{p}
-    + \frac{1}{\Reynolds} \nabla^2 \bs{v}
-    + \bs{f},
-    \qquad \qquad \forall (\bs{x}, t) \in \Omega \times [0, T],
-    \end{gather*}
-  * \f]
-  * where \f$ \bs{v} \f$ is the velocity,
-  * \f$ \rho_0 \f$ the density of the reference state,
-  * \f$ \bar{p} \f$ the modified pressure,
-  * \f$ \Reynolds \f$ the Reynolds number and
-  * \f$ \bs{f} \f$ the body force,
-  */
-  hydrodynamic,
-
-  /*!
-   * @brief The heat convection-difussion equation.
-   * @details Given by
-  * \f[
-  * \begin{gather*}
-    \pd{\vartheta}{t} + \bs{v} \cdot \nabla \vartheta
-    =
-    \frac{1}{\Peclet} \nabla^2 \vartheta + r, \qquad \qquad \forall (\bs{x}, t) \in \Omega \times [0, T],
-    \end{gather*}
-  * \f]
-  * where \f$ \bs{v} \f$ is the velocity,
-  * \f$ \vartheta \f$ the temperature,
-  * \f$ \Peclet \f$ the Peclet number and
-  * \f$ r \f$ the internal heat generation.
-  */
-  heat_convection_diffusion,
-
-  /*!
-   * @brief Boussinesq-approximation in an interial frame of reference.
-   * @details Given by
-  * \f[
-  * \begin{gather*}
-    \nabla \cdot \bs{v}
-    = 0, \quad
-      \pd{\bs{v}}{t} + \bs{v} \cdot (\nabla \otimes \bs{v})
-    = - \nabla \bar{p}
-    + \sqrt{\frac{\Prandtl}{\Rayleigh}} \nabla^2 \bs{v}
-    - \vartheta \bs{g},
-    \\
-    \pd{\vartheta}{t} + \bs{v} \cdot \nabla \vartheta
-    =
-    \frac{1}{\sqrt{\Rayleigh \Prandtl}} \nabla^2 \vartheta, \qquad \qquad \forall (\bs{x}, t) \in \Omega \times [0, T],
-    \end{gather*}
-  * \f]
-  * where \f$ \bs{v} \f$ is the velocity,
-  * \f$ \bar{p} \f$ the modified pressure,
-  * \f$ \Prandtl \f$ the Prandtl number,
-  * \f$ \Rayleigh \f$ the Rayleigh number and
-  * \f$ \vartheta \f$ the temperature.
-  * @note If specifified, body forces and internal heat generation are
-  * also considered.
-  */
-  boussinesq,
-
-  /*!
-   * @brief Boussinesq-approximation in an rotating frame of reference.
-   * @details Given by
-  * \f[
-  * \begin{gather*}
-    \nabla \cdot \bs{v}
-    = 0, \quad
-      \pd{\bs{v}}{t} + \bs{v} \cdot (\nabla \otimes \bs{v})
-      + \frac{2}{\Ekman} \bs{\Omega} \times \bs{v}
-    = -\dfrac{1}{\Ekman} \nabla \bar{p}
-    + \nabla^2 \bs{v}
-    - \frac{\Rayleigh}{\Prandtl} \vartheta \bs{g},
-    \\
-    \pd{\vartheta}{t} + \bs{v} \cdot \nabla \vartheta
-    =
-    \frac{1}{\Prandtl} \nabla^2 \vartheta, \qquad \qquad \forall (\bs{x}, t) \in \Omega \times [0, T],
-    \end{gather*}
-  * \f]
-  * where \f$ \bs{v} \f$ is the velocity,
-  * \f$ \bs{\Omega} \f$ the angular velocity,
-  * \f$ \bar{p} \f$ the modified pressure,
-  * \f$ \Ekman \f$ the Ekman number,
-  * \f$ \Rayleigh \f$ the Rayleigh number,
-  * \f$ \Prandtl \f$ the Prandtl number,
-  * \f$ \vartheta \f$ the temperature.
-  * @note If specifified, body forces and internal heat generation are
-  * also considered.
-  */
-  rotating_boussinesq,
-
-  /*!
-   * @brief Boussinesq-approximation with electromagnetic forces in an rotating frame of reference.
-   * @details Given by
-  * \f[
-  * \begin{gather*}
-    \nabla \cdot \bs{v}
-    = 0, \quad
-      \pd{\bs{v}}{t} + \bs{v} \cdot (\nabla \otimes \bs{v})
-      + \frac{2}{\Ekman} \bs{\Omega} \times \bs{v}
-    = -\dfrac{1}{\Ekman} \nabla \bar{p}
-    + \nabla^2 \bs{v}
-    - \frac{\Rayleigh}{\Prandtl} \vartheta \bs{g}
-    + \frac{1}{\magPrandtl}(\nabla \times \bs{B}) \times \bs{B},
-    \\
-    \pd{\vartheta}{t} + \bs{v} \cdot \nabla \vartheta
-    =
-    \frac{1}{\Prandtl} \nabla^2 \vartheta, \\
-    \nabla \cdot \bs{B} = 0, \quad
-    \pd{\bs{B}}{t} = \nabla \times ( \bs{v} \times \bs{B} ) + \frac{1}{\magPrandtl} \nabla \times ( \nabla \times \bs{B} ),    \qquad \qquad \forall (\bs{x}, t) \in \Omega \times [0, T],
-    \end{gather*}
-  * \f]
-  * where \f$ \bs{v} \f$ is the velocity,
-  * \f$ \bs{\Omega} \f$ the angular velocity,
-  * \f$ \Ekman \f$ the Ekman number,
-  * \f$ \bar{p} \f$ the modified pressure,
-  * \f$ \Rayleigh \f$ the Rayleigh number,
-  * \f$ \Prandtl \f$ the Prandtl number,
-  * \f$ \vartheta \f$ the temperature,
-  * \f$ \bs{B} \f$ the magnetic field and
-  * \f$ \magPrandtl \f$ the magnetic Prandtl number.
-  * @note If specifified, body forces and internal heat generation are
-  * also considered.
-  */
-  rotating_magnetohydrodynamic
-};
-
-
-
-/*!
- * @brief Enumeration for incremental pressure-correction scheme types.
- */
-enum class PressureCorrectionScheme
-{
-  /*!
-   * @brief Standard incremental pressure-correction scheme.
-   * @details The pressure update is given by
-   * \f[
-   * p^{k} = p^{k-1} + \phi^{k}
-   * \f]
-   * @note See @ref NavierStokesProjection for a complete explanation
-   * of the incremental pressure-correction scheme.
-   */
-  standard,
-
-  /*!
-   * @brief Rotational incremental pressure-correction scheme.
-   * @details The pressure update is given by
-   * \f[
-   * p^{k} = p^{k-1} + \phi^{k} - \nu \nabla \cdot \bs{u}^k
-   * \f]
-   * @note See @ref NavierStokesProjection for a complete explanation
-   * of the incremental pressure-correction scheme.
-   */
-  rotational
-};
-
-
-
-/*!
- * @brief Enumeration for the weak form of the non-linear convective term.
- * @attention These definitions are the ones I see the most in the literature.
- * Nonetheless Volker John and Helene Dallmann define the skew-symmetric
- * and the divergence form differently.
- */
-enum class ConvectiveTermWeakForm
-{
-  /*!
-   * @brief The standard form.
-   * @details Given by
-   * \f[
-   * \int_\Omega \bs{w} \cdot [ \bs{v} \cdot ( \nabla \otimes \bs{v})] \dint{v}
-   * \f]
-   * where \f$ \bs{v} \f$ and \f$ \bs{w} \f$ are the velocity and the
-   * test function respectively.
-   */
-  standard,
-
-  /*!
-   * @brief The skew-symmetric form.
-   * @details Given by
-   * \f[
-   * \int_\Omega \bs{w} \cdot [ \bs{v} \cdot ( \nabla \otimes \bs{v}) +
-   * (\nabla \cdot \bs{v}) \bs{v}] \dint{v}
-   * \f]
-   * where \f$ \bs{v} \f$ and \f$ \bs{w} \f$ are the velocity and the
-   * test function respectively.
-   */
-  skewsymmetric,
-
-  /*!
-   * @brief The divergence form.
-   * @details Given by
-   * \f[
-   * \int_\Omega \bs{w} \cdot [ \bs{v} \cdot ( \nabla \otimes \bs{v}) +
-   * \frac{1}{2}(\nabla \cdot \bs{v}) \bs{v}] \dint{v}
-   * \f]
-   * where \f$ \bs{v} \f$ and \f$ \bs{w} \f$ are the velocity and the
-   * test function respectively.
-   */
-  divergence,
-
-  /*!
-   * @brief The rotational form.
-   * @details Given by
-   * \f[
-   * \int_\Omega \bs{w} \cdot [ ( \nabla \times\bs{v}) \times \bs{v}] \dint{v}
-   * \f]
-   * where \f$ \bs{v} \f$ and \f$ \bs{w} \f$ are the velocity and the
-   * test function respectively.
-   * @note This form modifies the pressure, *i. e.*,
-   * \f[
-   * \bar{p} = p + \frac{1}{2} \bs{v} \cdot \bs{v}.
-   * \f]
-   */
-  rotational
-};
-
-
-
-/*!
- * @brief Enumeration for the type of the temporal discretization of the
- * non-linear convective term.
- */
-enum class ConvectiveTermTimeDiscretization
-{
-  /*!
-   * @brief Semi-implicit treatment in time of the convective term, *i. e.*,
-   * \f[
-   * \bs{v}^{\star,k} \cdot
-   * (\nabla \otimes\bs{\phi}^{k}),
-   * \f]
-   * where \f$\bs{v}^{\star, k}\f$ is the second order Taylor extrapolation
-   * of the velocity at \f$t^k \f$ and \f$ \bs{\phi}^{k} \f$ is the
-   * field being transported by the velocity.
-   * @note The semi-implicit treatment for non-standard weak forms of
-   * the convective terms is given
-   * by
-   * \f[
-   * \bs{v}^{k} \cdot ( \nabla \otimes \bs{v}^{\star,k}) +
-   * \frac{1}{2}(\nabla \cdot \bs{v}^{\star,k}) \bs{v}^{k},
-   * \qquad
-   * \bs{v}^{k} \cdot ( \nabla \otimes \bs{v}^{\star,k}) +
-   * (\nabla \cdot \bs{v}^{\star,k}) \bs{v}^{k},
-   * \qquad
-   * ( \nabla \times\bs{v}^{\star,k}) \times \bs{v}^{k},
-   * \f]
-   * for the skew-symmetric, divergence and rotational form respectively.
-   */
-  semi_implicit,
-
-  /*!
-   * @brief Explicit treatment in time of the convective term, *i. e.*,
-   * \f[
-   * \sum_{j = 1} \beta_j
-   * \bs{v}^{k-j} \cdot
-   * (\nabla \otimes\bs{\phi}^{k-j}),
-   * \f]
-   * where \f$ \beta_j \f$ are time discretization coefficientes,
-   * \f$\bs{v}^{k-j}\f$ the previous values of the velocity field and
-   * \f$ \bs{\phi}^{k-j} \f$ the previous values of the
-   * field being transported by the velocity.
-   * @attention Why fully_explicit and not just explicit?
-   */
-  fully_explicit
-};
-
-/*!
- * @brief Enumeration for the type of the preconditioner to be used.
- */
-enum class PreconditionerType
-{
-  /*!
-   * @brief Incomplete LU decomposition preconditioning.
-   */
-  ILU,
-
-  /*!
-   * @brief Geometric multigrid preconditioning.
-   *
-   * @attention This is not implemented yet but may be helpful once the
-   * MatrixFree framework is used.
-   */
-  GMG,
-
-  /*!
-   * @brief Algebraic multigrid preconditioning.
-   */
-  AMG,
-
-  /*!
-   * @brief Jacobi preconditioning.
-   */
-  Jacobi,
-
-  /*!
-   * @brief Symmetric Sucessive overrelaxation preconditioning. The system
-   * matrix must be symmetric to apply this preconditioner.
-   */
-  SSOR
-
-};
 
 /*!
  * @struct SpatialDiscretizationParameters
@@ -517,419 +176,6 @@ struct OutputControlParameters
 template<typename Stream>
 Stream& operator<<(Stream &stream, const OutputControlParameters &prm);
 
-
-/*!
- * @struct PreconditionerParametersBase
- *
- * @brief A structure from which all other preconditioner parameters structures
- * are derived.
- */
-struct PreconditionBaseParameters
-{
-  /*!
-   * Constructor which sets up the parameters with default values.
-   */
-  PreconditionBaseParameters(const std::string        &name = "ILU",
-                             const PreconditionerType &type = PreconditionerType::ILU);
-
-  /*!
-   * @brief Static method which declares the associated parameter to the
-   * ParameterHandler object @p prm.
-   */
-  static void declare_parameters(ParameterHandler &prm);
-
-  /*!
-   * @brief Method which parses the only the type parameters of
-   * ::PreconditionBaseParameters from the ParameterHandler object @p prm.
-   */
-  static PreconditionerType parse_preconditioner_type(const ParameterHandler &prm);
-
-  /*!
-   * @brief Method which parses the parameters of ::PreconditionBaseParameters
-   * from the ParameterHandler object @p prm.
-   */
-  void parse_parameters(const ParameterHandler &prm);
-
-  /*!
-   * @brief The type of the preconditioner used for solving the linear system.
-   */
-  PreconditionerType preconditioner_type;
-
-  /*!
-   * @brief The name of the preconditioner.
-   */
-  std::string       preconditioner_name;
-};
-
-
-/*!
- * @struct PreconditionRelaxationParameters
- *
- * @brief A structure containing all parameters relevant for Jacobi
- * preconditioning.
- */
-struct PreconditionRelaxationParameters : PreconditionBaseParameters
-{
-  /*!
-   * Constructor which sets up the parameters with default values.
-   */
-  PreconditionRelaxationParameters();
-
-  /*!
-   * @brief Static method which declares the associated parameter to the
-   * ParameterHandler object @p prm.
-   */
-  static void declare_parameters(ParameterHandler &prm);
-
-  /*!
-   * @brief Method which parses the parameters of the ILU preconditioner from
-   * the ParameterHandler object @p prm.
-   */
-  void parse_parameters(const ParameterHandler &prm);
-
-  /*!
-   * @brief Method forwarding parameters to a stream object.
-   *
-   * @details This method does not add a `std::endl` to the stream at the end.
-   *
-   */
-  template<typename Stream>
-  friend Stream& operator<<(Stream &stream, const PreconditionRelaxationParameters &prm);
-
-  /*!
-   * @brief The relaxation parameter in the preconditioner.
-   *
-   * @attention For a Jacobi preconditioner, this parameter is only meaningful if the library is
-   * compiled using the Trilinos linear algebra package.
-   *
-   */
-  double        omega;
-
-  /*!
-   * @brief A parameter specifying how large the overlap of the local matrix
-   * portions on each processor in a parallel application should be.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package and if an SOR or SSOR
-   * preconditioner is used.
-   *
-   */
-  unsigned int  overlap;
-
-
-  /*!
-   * @brief A parameter specifying how many times the given preconditioner
-   * should be applied at each call.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   *
-   */
-  unsigned int  n_sweeps;
-};
-
-using PreconditionJacobiParameters = PreconditionRelaxationParameters;
-
-using PreconditionSSORParameters = PreconditionRelaxationParameters;
-
-/*!
- * @brief Method forwarding parameters to a stream object.
- *
- * @details This method does not add a `std::endl` to the stream at the end.
- */
-template<typename Stream>
-Stream& operator<<(Stream &stream, const PreconditionRelaxationParameters &prm);
-
-/*!
- * @struct PreconditionILUParameters
- *
- * @brief A structure containing all parameters relevant for the computation of
- * an incomplete LU preconditioner.
- *
- */
-struct PreconditionILUParameters : PreconditionBaseParameters
-{
-  /*!
-   * Constructor which sets up the parameters with default values.
-   */
-  PreconditionILUParameters();
-
-  /*!
-   * @brief Static method which declares the associated parameter to the
-   * ParameterHandler object @p prm.
-   */
-  static void declare_parameters(ParameterHandler &prm);
-
-  /*!
-   * @brief Method which parses the parameters of the ILU preconditioner from
-   * the ParameterHandler object @p prm.
-   */
-  void parse_parameters(const ParameterHandler &prm);
-
-  /*!
-   * @brief Method forwarding parameters to a stream object.
-   *
-   * @details This method does not add a `std::endl` to the stream at the end.
-   *
-   */
-  template<typename Stream>
-  friend Stream& operator<<(Stream &stream, const PreconditionILUParameters &prm);
-
-  /*!
-   * @brief Relative tolerance used for the perturbation of the diagonal entries
-   * of the diagonal matrix which meet yield a better preconditioning. For the
-   * relative tolerance \f$\beta\f$ the condition \f$\beta\geq 1\f$ must hold.
-   * The suggested value is \f$\beta\approx 1.01\f$.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   *
-   */
-  double        relative_tolerance;
-
-  /*!
-   * @brief Absolute tolerance used for the perturbation of the diagonal entries
-   * of the diagonal matrix which meet yield a better preconditioning. For the
-   * relative tolerance \f$\alpha\f$ the condition \f$\alpha>0\f$ must hold. The
-   * suggested value is \f$10^{-5}\alpha\leq 10^{-2}\f$.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   */
-  double        absolute_tolerance;
-
-  /*!
-   * @brief Fill-in level of the ILU. The ILU uses only the entries of the
-   * sparsity pattern of the matrix \f$ A^k\f$ where \f$ k\f$ is the fill-in
-   * level.
-   */
-  unsigned int  fill;
-
-  /*!
-   * @brief Overlap between the different MPI processes used for computing
-   * the ILU.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   */
-  unsigned int  overlap;
-};
-
-/*!
- * @brief Method forwarding parameters to a stream object.
- *
- * @details This method does not add a `std::endl` to the stream at the end.
- */
-template<typename Stream>
-Stream& operator<<(Stream &stream, const PreconditionILUParameters &prm);
-
-
-/*!
- * @struct PreconditionAMGParameters
- *
- * @brief A structure containing the common parameters used in AMG
- * preconditioning.
- */
-struct PreconditionAMGParameters : PreconditionBaseParameters
-{
-  /*!
-   * Constructor which sets up the parameters with default values.
-   */
-  PreconditionAMGParameters();
-
-  /*!
-   * @brief Static method which declares the associated parameter to the
-   * ParameterHandler object @p prm.
-   */
-  static void declare_parameters(ParameterHandler &prm);
-
-  /*!
-   * @brief Method which parses the parameters of the AMG preconditioner from
-   * the ParameterHandler object @p prm.
-   */
-  void parse_parameters(const ParameterHandler &prm);
-
-  /*!
-   * @brief Method forwarding parameters to a stream object.
-   *
-   * @details This method does not add a `std::endl` to the stream at the end.
-   *
-   */
-  template<typename Stream>
-  friend Stream& operator<<(Stream &stream, const PreconditionAMGParameters &prm);
-
-  /*
-   * @brief A control parameter for AMG preconditioning using the PETSc library.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the PETSc linear algebra package.
-   *
-   */
-  double strong_threshold;
-
-  /*
-   * @brief A control parameter to control whether smoothed aggregation or
-   * non-smoothed aggregration should be applied.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   *
-   */
-  bool  elliptic;
-
-  /*
-   * @brief A control parameter to control whether higher order finite elements
-   * are used.
-   *
-   * @details The parameters cannot be specified through the parameter file but
-   * is set according to the polynomial degree of the finite element.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   *
-   */
-  bool  higher_order_elements;
-
-  /*
-   * @brief A control parameter to control how many multigrid cycles should be
-   * performed.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   *
-   */
-  unsigned int  n_cycles;
-
-  /*
-   * @brief A control parameter for AMG preconditioning using the Trilinos
-   * library to control how the coarsening should be performed.
-   *
-   * @attention This parameter is only meaningful if the library is compiled
-   * using the Trilinos linear algebra package.
-   *
-   */
-  double  aggregation_threshold;
-
-};
-
-
-/*!
- * @brief Method forwarding parameters to a stream object.
- *
- * @details This method does not add a `std::endl` to the stream at the end.
- */
-template<typename Stream>
-Stream& operator<<(Stream &stream, const PreconditionAMGParameters &prm);
-
-
-/*!
- * @struct LinearSolverParameters
- *
- * @brief A structure containing all parameters relevant for the solution of
- * linear systems using a Krylov subspace method.
- *
- * @todo Proper initiation of the solver_name string without constructor
- * ambiguity. The string is included in order for the stream to print
- * also the names of the solvers to the terminal. As of now one can not
- * tell one from the other (On the prm file they are properly differentiated)
- */
-struct LinearSolverParameters
-{
-  /*!
-   * Constructor which sets up the parameters with default values.
-   */
-  LinearSolverParameters(const std::string &name = "default");
-
-  /*!
-   * @brief Static method which declares the associated parameter to the
-   * ParameterHandler object @p prm.
-   */
-  static void declare_parameters(ParameterHandler &prm);
-
-  /*!
-   * @brief Method which parses the parameters of the linear solver from
-   * the ParameterHandler object @p prm.
-   */
-  void parse_parameters(ParameterHandler &prm);
-
-  /*!
-   * @brief Method forwarding parameters to a stream object.
-   *
-   * @details This method does not add a `std::endl` to the stream at the end.
-   *
-   */
-  template<typename Stream>
-  friend Stream& operator<<(Stream &stream, const LinearSolverParameters &prm);
-
-  /*!
-   * @brief Relative tolerance.
-   *
-   * @details This parameter specifies the desired relative reduction of the
-   * initial residual of the linear system. The initial residual is given by
-   *
-   * \f[
-   *    r_0=|\!| \mathbf{A}\cdot\mathbf{x}_0-\mathbf{b}|\!|_2\,.
-   * \f]
-   *
-   * If the relative tolerance is applied, the final solution of the linear
-   * system is such that
-   *
-   * \f[
-   *    |\!| \mathbf{A}\cdot\mathbf{x}-\mathbf{b}|\!|_2\leq\mathrm{tol}\,r_0\,.
-   * \f]
-   *
-   */
-  double        relative_tolerance;
-
-  /*!
-   * @brief Absolute tolerance.
-   *
-   * @details This parameter specifies the desired absolute value of the
-   * residual of the linear system.
-   *
-   * If the absolute tolerance is applied, the final solution of the linear
-   * system is such that
-   *
-   * \f[
-   *    |\!| \mathbf{A}\cdot\mathbf{x}-\mathbf{b}|\!|_2\leq\mathrm{tol}\,.
-   * \f]
-   *
-   */
-  double        absolute_tolerance;
-
-  /*!
-   * @brief Maximum number of iterations to be performed by the linear solver.
-   *
-   * @details If the tolerance is fulfilled within the maximum number of
-   * iterations, an error is thrown and the simulation is aborted. In this case,
-   * the preconditioner might require some optimization.
-   *
-   */
-  unsigned int  n_maximum_iterations;
-
-  /*!
-   * @brief Pointer to the parameter of the preconditioners
-   */
-  std::shared_ptr<PreconditionBaseParameters> preconditioner_parameters_ptr;
-
-private:
-
-  /*!
-   * @brief The name of the solver.
-   */
-  std::string   solver_name;
-};
-
-
-
-/*!
- * @brief Method forwarding parameters to a stream object.
- *
- * @details This method does not add a `std::endl` to the stream at the end.
- *
- */
-template<typename Stream>
-Stream& operator<<(Stream &stream, const LinearSolverParameters &prm);
 
 
 
@@ -1201,7 +447,6 @@ template<typename Stream>
 Stream& operator<<(Stream &stream, const NavierStokesParameters &prm);
 
 
-
 /*!
  * @struct HeatEquationParameters
  *
@@ -1288,6 +533,7 @@ struct HeatEquationParameters
  */
 template<typename Stream>
 Stream& operator<<(Stream &stream, const HeatEquationParameters &prm);
+
 
 
 
@@ -1622,8 +868,6 @@ private:
  */
 template<typename Stream>
 Stream& operator<<(Stream &stream, const ProblemParameters &prm);
-
-
 
 } // namespace RunTimeParameters
